move logCoreAnalyzer into its own folder
